+++ source/renderer/zettlr-rendereripc.js
@@ -590,15 +590,6 @@ class ZettlrRendererIPC {
         this._app.getEditor().zoom(-1)
         break
 
-      // Updater
-      case 'update-check':
-        this.send('update-check')
-        break
-
-      case 'update-available':
-        this._app.getBody().displayUpdate(cnt)
-        break
-
       // About dialog
       case 'display-about':
         this._app.getBody().displayAbout()